Major Change: updated transferWindow
The issue that remains is the impossibility of the window in the other
side of the connection to be automatically close when the user
interrupt the trasnfer.

+++ Movex.FTP/DTransfer.cs
@@ -114,7 +114,7 @@ namespace Movex.FTP
         }
         public float GetTransferPerc()
         {
-            return ((GetTransfered() / mToTransfer) * 100);
+            return (((float)GetTransfered() / mToTransfer) * 100);
         }
         public string GetFrom()
         {
